requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cryptocompare==0.7.6
frozenlist==1.4.1
idna==3.7
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
peewee==3.17.5
peewee-async==0.10.0
peewee-migrate==1.12.2
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
pytz==2024.1
redis==5.0.4
requests==2.32.2
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
